+++ src/main/java/me/roinujnosde/titansbattle/challenges/ChallengeRequest.java
@@ -3,6 +3,8 @@ package me.roinujnosde.titansbattle.challenges;
 import me.roinujnosde.titansbattle.types.Warrior;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.Objects;
+
 public abstract class ChallengeRequest<T> {
 
     protected final Challenge challenge;
@@ -17,7 +19,24 @@ public abstract class ChallengeRequest<T> {
 
     public abstract boolean isInvited(@NotNull Warrior warrior);
 
-    public @NotNull Challenge getArena() {
+    public @NotNull Challenge getChallenge() {
         return challenge;
+    }
+
+    public T getChallenger() {
+        return challenger;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        ChallengeRequest<?> that = (ChallengeRequest<?>) o;
+        return challenge.equals(that.challenge) && challenger.equals(that.challenger) && challenged.equals(that.challenged);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(challenge, challenger, challenged);
     }
 }
