+++ src/main/java/me/roinujnosde/titansbattle/events/GameStartEvent.java
@@ -23,7 +23,7 @@
  */
 package me.roinujnosde.titansbattle.events;
 
-import me.roinujnosde.titansbattle.games.Game;
+import me.roinujnosde.titansbattle.BaseGame;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
@@ -37,17 +37,16 @@ public class GameStartEvent extends Event implements Cancellable {
     
     private static final HandlerList HANDLERS = new HandlerList();
     private boolean cancelled = false;
-    private final Game game;
+    private final BaseGame game;
 
-    public GameStartEvent(@NotNull Game game) {
+    public GameStartEvent(@NotNull BaseGame game) {
         this.game = game;
     }
     
     /**
-     * Returns the Game
      * @return the Game
      */
-    public Game getGame() {
+    public BaseGame getGame() {
         return game;
     }
