+++ src/main/java/me/roinujnosde/titansbattle/games/Game.java
@@ -2,111 +2,36 @@ package me.roinujnosde.titansbattle.games;
 
 import me.roinujnosde.titansbattle.BaseGame;
 import me.roinujnosde.titansbattle.TitansBattle;
-import me.roinujnosde.titansbattle.events.*;
+import me.roinujnosde.titansbattle.events.LobbyStartEvent;
 import me.roinujnosde.titansbattle.managers.DatabaseManager;
-import me.roinujnosde.titansbattle.managers.GameManager;
-import me.roinujnosde.titansbattle.managers.GroupManager;
 import me.roinujnosde.titansbattle.types.GameConfiguration;
 import me.roinujnosde.titansbattle.types.Group;
 import me.roinujnosde.titansbattle.types.Kit;
 import me.roinujnosde.titansbattle.types.Warrior;
-import me.roinujnosde.titansbattle.utils.SoundUtils;
-import net.md_5.bungee.api.ChatMessageType;
-import net.md_5.bungee.api.chat.TextComponent;
 import org.bukkit.Bukkit;
-import org.bukkit.ChatColor;
-import org.bukkit.Location;
-import org.bukkit.Material;
-import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
-import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.scheduler.BukkitRunnable;
 import org.bukkit.scheduler.BukkitTask;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.text.MessageFormat;
-import java.util.*;
-import java.util.function.Consumer;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import static me.roinujnosde.titansbattle.utils.SoundUtils.Type.*;
 
 public abstract class Game extends BaseGame {
 
-    protected final GameManager gameManager;
     protected final DatabaseManager databaseManager;
     private final GameConfiguration config;
-    protected boolean lobby;
-    private final List<BukkitTask> tasks = new ArrayList<>();
     protected boolean battle = false;
 
     public Game(TitansBattle plugin, GameConfiguration config) {
         super(plugin);
         this.config = config;
         this.databaseManager = plugin.getDatabaseManager();
-        gameManager = plugin.getGameManager();
     }
-
-    @Override
-    protected boolean canJoin(@NotNull Warrior warrior) {
-        Player player = warrior.toOnlinePlayer();
-        String reason = null;
-        if (player == null) {
-            plugin.debug(String.format("canJoin() -> player %s %s == null", warrior.getName(), warrior.getUniqueId()), false);
-            return false;
-        }
-        if (!isLobby()) {
-            reason = plugin.getLang("game_is_happening", this);
-            plugin.debug("happening");
-        }
-        if (isParticipant(warrior)) {
-            reason = plugin.getLang("already-joined", this);
-            plugin.debug("already in");
-        }
-        if (participants.size() >= getConfig().getMaximumPlayers() && getConfig().getMaximumPlayers() > 0) {
-            reason = plugin.getLang("maximum-players", this);
-            plugin.debug("max players");
-        }
-        if (config.isGroupMode()) {
-            if (warrior.getGroup() == null) {
-                reason = plugin.getLang("not_in_a_group", this);
-                plugin.debug("not in group");
-            }
-            if (getGroupParticipants().size() >= getConfig().getMaximumGroups() && getConfig().getMaximumGroups() > 0) {
-                reason = plugin.getLang("maximum-groups", this);
-                plugin.debug("max groups");
-            }
-            Integer amountOfPlayers = getGroupParticipants().getOrDefault(warrior.getGroup(), 0);
-            if (amountOfPlayers >= getConfig().getMaximumPlayersPerGroup() &&
-                    getConfig().getMaximumPlayersPerGroup() > 0) {
-                reason = plugin.getLang("maximum-players-per-group", this);
-                plugin.debug("max per group");
-            }
-        }
-        if (getConfig().isUseKits() && Kit.inventoryHasItems(player)) {
-            reason = plugin.getLang("clear-your-inventory", this);
-            plugin.debug("clear inv");
-        }
-
-        PlayerJoinGameEvent event = new PlayerJoinGameEvent(player, this);
-        Bukkit.getPluginManager().callEvent(event);
-        plugin.debug("cancel: " + event.isCancelled());
-        plugin.debug("reason: >>>" + reason + "<<<");
-        if (reason != null) {
-            player.sendMessage(reason);
-        }
-        return reason == null && !event.isCancelled();
-    }
-
-    public abstract boolean isInBattle(@NotNull Warrior warrior);
-
-    public boolean isParticipant(@NotNull Warrior warrior) {
-        return participants.contains(warrior);
-    }
-
 
     public boolean shouldClearDropsOnDeath(@NotNull Warrior warrior) {
         return isParticipant(warrior) && config.isClearItemsOnDeath();
@@ -114,31 +39,6 @@ public abstract class Game extends BaseGame {
 
     public boolean shouldKeepInventoryOnDeath(@NotNull Warrior warrior) {
         return false;
-    }
-
-    protected void processPlayerExit(@NotNull Warrior warrior) {
-        if (!isParticipant(warrior)) {
-            return;
-        }
-        Player player = warrior.toOnlinePlayer();
-        if (player != null) {
-            teleport(warrior, getConfig().getExit());
-            PlayerExitGameEvent event = new PlayerExitGameEvent(player, this);
-            Bukkit.getPluginManager().callEvent(event);
-        }
-        participants.remove(warrior);
-        Group group = warrior.getGroup();
-        if (!isLobby()) {
-            runCommandsAfterBattle(Collections.singletonList(warrior));
-            processRemainingPlayers(warrior);
-            //last participant
-            if (config.isGroupMode() && group != null && !getGroupParticipants().containsKey(group)) {
-                gameManager.broadcastKey("group_defeated", this, group.getName());
-                Bukkit.getPluginManager().callEvent(new GroupDefeatedEvent(group, warrior.toOnlinePlayer()));
-                group.getData().increaseDefeats(getConfig().getName());
-            }
-            sendRemainingOpponentsCount();
-        }
     }
 
     /**
@@ -162,116 +62,20 @@ public abstract class Game extends BaseGame {
         return killerPlayer;
     }
 
-    protected abstract void processRemainingPlayers(@NotNull Warrior warrior);
-
-    public void onLeave(@NotNull Warrior warrior) {
-        if (!isParticipant(warrior)) {
-            return;
-        }
-        if (getConfig().isUseKits()) {
-            Kit.clearInventory(warrior.toOnlinePlayer());
-        }
-        Player player = Objects.requireNonNull(warrior.toOnlinePlayer());
-        player.sendMessage(plugin.getLang("you-have-left", this));
-        SoundUtils.playSound(LEAVE_GAME, plugin.getConfig(), player);
-        processPlayerExit(warrior);
-    }
-
-    public void onDisconnect(@NotNull Warrior warrior) {
-        if (!isParticipant(warrior)) {
-            return;
-        }
-        if (config.isUseKits()) {
-            plugin.getConfigManager().getClearInventory().add(warrior.getUniqueId());
-        }
-        plugin.getConfigManager().getRespawn().add(warrior.getUniqueId());
-        plugin.getConfigManager().save();
-        processPlayerExit(warrior);
-    }
-
-    public void onRespawn(@NotNull Warrior warrior) {
-        if (casualties.contains(warrior) && !casualtiesWatching.contains(warrior)) {
-            teleport(warrior, getConfig().getWatchroom());
-            casualtiesWatching.add(warrior);
-        }
-    }
-
-    public void onDeath(@NotNull Warrior victim, @Nullable Warrior killer) {
-        if (!isParticipant(victim)) {
-            return;
-        }
-        if (!isLobby()) {
-            ParticipantDeathEvent event = new ParticipantDeathEvent(victim, killer);
-            Bukkit.getPluginManager().callEvent(event);
-            String gameName = getConfig().getName();
-            casualties.add(victim);
-            if (config.isGroupMode()) {
-                victim.sendMessage(plugin.getLang("watch_to_the_end", this));
-            }
-            if (killer != null) {
-                killer.increaseKills(gameName);
-                increaseKills(killer);
-            }
-            victim.increaseDeaths(gameName);
-            playDeathSound(victim);
-        }
-        broadcastDeathMessage(victim, killer);
-        processPlayerExit(victim);
-    }
-
-    @SuppressWarnings("deprecation")
-    protected void broadcastDeathMessage(@NotNull Warrior victim, @Nullable Warrior killer) {
-        if (killer == null) {
-            gameManager.broadcastKey("died_by_himself", this, victim.getName());
-        } else {
-            ItemStack itemInHand = Objects.requireNonNull(killer.toOnlinePlayer()).getItemInHand();
-            String weaponName = plugin.getLang("fist", this);
-            if (itemInHand != null && itemInHand.getType() != Material.AIR) {
-                ItemMeta itemMeta = itemInHand.getItemMeta();
-                if (itemMeta != null && itemMeta.hasDisplayName()) {
-                    weaponName = itemMeta.getDisplayName();
-                }
-            }
-            gameManager.broadcastKey("killed_by", this, victim.getName(),
-                    killsCount.getOrDefault(victim, 0), killer.getName(), killsCount.get(killer), weaponName);
-        }
-    }
-
-    protected abstract void onLobbyEnd();
-
-    protected boolean canStartBattle() {
-        GameStartEvent event = new GameStartEvent(this);
-        Bukkit.getPluginManager().callEvent(event);
-        if (event.isCancelled()) {
-            gameManager.broadcastKey("cancelled", this, "CONSOLE");
-            return false;
-        }
-        if (getParticipants().size() < getConfig().getMinimumPlayers()) {
-            gameManager.broadcastKey("not_enough_participants", this);
-            return false;
-        }
-        if (getConfig().isGroupMode()) {
-            if (getGroupParticipants().size() < getConfig().getMinimumGroups()) {
-                gameManager.broadcastKey("not_enough_participants", this);
-                return false;
-            }
-        }
-        return true;
-    }
-
+    @Override
     public void start() {
         LobbyStartEvent event = new LobbyStartEvent(this);
         Bukkit.getPluginManager().callEvent(event);
         if (event.isCancelled()) {
             return;
         }
-        gameManager.setCurrentGame(this);
         if (getConfig().isGroupMode() && plugin.getGroupManager() == null) {
             throw new IllegalStateException("You cannot start a group based game without a supported Groups plugin!");
         }
         if (!getConfig().locationsSet()) {
             throw new IllegalStateException("You didn't set all locations!");
         }
+        gameManager.setCurrentGame(this);
         lobby = true;
         Integer interval = config.getAnnouncementStartingInterval();
         BukkitTask lobbyTask = new LobbyAnnouncementTask(config.getAnnouncementStartingTimes(), interval)
@@ -279,19 +83,18 @@ public abstract class Game extends BaseGame {
         addTask(lobbyTask);
     }
 
-    protected void addTask(@NotNull BukkitTask task) {
-        tasks.add(task);
-    }
-
-    protected void killTasks() {
-        tasks.forEach(BukkitTask::cancel);
-        tasks.clear();
-        plugin.getTaskManager().killAllTasks();
-    }
-
-    public void cancel(@NotNull CommandSender sender) {
-        gameManager.broadcastKey("cancelled", this, sender.getName());
-        finish(true);
+    public void finish(boolean cancelled) {
+        teleportAll(getConfig().getExit());
+        killTasks();
+        runCommandsAfterBattle(getParticipants());
+        if (getConfig().isUseKits()) {
+            getPlayerParticipantsStream().forEach(Kit::clearInventory);
+        }
+        gameManager.setCurrentGame(null);
+        if (!cancelled) {
+            processWinners();
+        }
+        plugin.getDatabaseManager().saveAll();
     }
 
     private void deleteGroups() {
@@ -304,8 +107,7 @@ public abstract class Game extends BaseGame {
                 }
             }
             if (deleted != 0) {
-                Bukkit.getServer().broadcastMessage(MessageFormat.format(plugin.getLang("deleted_groups", this),
-                        deleted));
+                broadcastKey("deleted_groups", deleted);
             }
         }
     }
@@ -334,146 +136,8 @@ public abstract class Game extends BaseGame {
         getConfig().getPrizes(prize).give(plugin, leaders, members);
     }
 
-    public void finish(boolean cancelled) {
-        teleportAll(getConfig().getExit());
-        killTasks();
-        runCommandsAfterBattle(getParticipants());
-        if (getConfig().isUseKits()) {
-            getPlayerParticipantsStream().forEach(Kit::clearInventory);
-        }
-        gameManager.setCurrentGame(null);
-        if (!cancelled) {
-            processWinners();
-        }
-        plugin.getDatabaseManager().saveAll();
-    }
-
     public @NotNull GameConfiguration getConfig() {
         return config;
-    }
-
-    public abstract @NotNull Collection<Warrior> getCurrentFighters();
-
-    public Map<Group, Integer> getGroupParticipants() {
-        Map<Group, Integer> groups = new HashMap<>();
-        for (Warrior w : participants) {
-            groups.compute(w.getGroup(), (g, i) -> i == null ? 1 : i + 1);
-        }
-        return groups;
-    }
-
-    public Set<Warrior> getCasualties() {
-        return casualties;
-    }
-
-    public HashMap<Warrior, Integer> getKillsCount() {
-        return killsCount;
-    }
-
-    protected void increaseKills(Warrior warrior) {
-        killsCount.compute(warrior, (p, i) -> i == null ? 1 : i + 1);
-    }
-
-
-
-    protected int getRemainingOpponents(@NotNull Player player) {
-        if (!getConfig().isGroupMode()) {
-            return getParticipants().size() - 1;
-        }
-        int opponents = 0;
-        Warrior warrior = plugin.getDatabaseManager().getWarrior(player);
-        for (Map.Entry<Group, Integer> entry : getGroupParticipants().entrySet()) {
-            Group group = entry.getKey();
-            if (group.equals(warrior.getGroup())) {
-                continue;
-            }
-            opponents += entry.getValue();
-        }
-        return opponents;
-    }
-
-    private void playDeathSound(@NotNull Warrior victim) {
-        Stream<Player> players = getPlayerParticipantsStream();
-        if (!getConfig().isGroupMode()) {
-            players.forEach(p -> SoundUtils.playSound(ENEMY_DEATH, plugin.getConfig(), p));
-            return;
-        }
-        GroupManager groupManager = plugin.getGroupManager();
-        if (groupManager == null) {
-            return;
-        }
-        Group victimGroup = groupManager.getGroup(victim.getUniqueId());
-        players.forEach(participant -> {
-            Group group = groupManager.getGroup(participant.getUniqueId());
-            if (group == null) {
-                return;
-            }
-            if (group.equals(victimGroup)) {
-                SoundUtils.playSound(ALLY_DEATH, plugin.getConfig(), participant);
-            } else {
-                SoundUtils.playSound(ENEMY_DEATH, plugin.getConfig(), participant);
-            }
-        });
-    }
-
-    protected void sendRemainingOpponentsCount() {
-        try {
-            getPlayerParticipantsStream().forEach(p -> {
-                int remaining = getRemainingOpponents(p);
-                if (remaining <= 0) {
-                    return;
-                }
-                p.spigot().sendMessage(ChatMessageType.ACTION_BAR, TextComponent.fromLegacyText(
-                        MessageFormat.format(plugin.getLang("action-bar-remaining-opponents", this),
-                                remaining)));
-            });
-        } catch (NoSuchMethodError ignored) {
-        }
-    }
-
-    protected void runCommandsBeforeBattle(@NotNull Collection<Warrior> warriors) {
-        runCommands(warriors, getConfig().getCommandsBeforeBattle());
-    }
-
-    protected void runCommandsAfterBattle(@NotNull Collection<Warrior> warriors) {
-        runCommands(warriors, getConfig().getCommandsAfterBattle());
-    }
-
-    private void runCommands(@NotNull Collection<Warrior> warriors, @Nullable Collection<String> commands) {
-        if (commands == null) return;
-        Consumer<Player> dispatchCommands = (player) -> {
-            for (String command : commands) {
-                Bukkit.dispatchCommand(Bukkit.getConsoleSender(), command.replace("%player%", player.getName()));
-            }
-        };
-        warriors.stream().map(Warrior::toOnlinePlayer).filter(Objects::nonNull).forEach(dispatchCommands);
-    }
-
-    protected void teleport(@NotNull Collection<Warrior> collection, @NotNull Location destination) {
-        collection.forEach(p -> teleport(p, destination));
-    }
-
-    protected void teleportAll(Location destination) {
-        getParticipants().forEach(player -> teleport(player, destination));
-    }
-
-    @SuppressWarnings("BooleanMethodIsAlwaysInverted")
-    public boolean isLobby() {
-        return lobby;
-    }
-
-    @Override
-    public int hashCode() {
-        return getConfig().getName().hashCode();
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (!(other instanceof Game)) {
-            return false;
-        }
-        Game otherGame = (Game) other;
-        return otherGame.getConfig().getName().equals(getConfig().getName());
     }
 
     protected void startPreparationTask() {
@@ -482,57 +146,10 @@ public abstract class Game extends BaseGame {
                 .runTaskTimer(plugin, 0L, 20L));
     }
 
-    protected class PreparationTimeTask extends BukkitRunnable {
-
-        @Override
-        public void run() {
-            Bukkit.getServer().broadcastMessage(plugin.getLang("preparation_over", Game.this));
-            runCommandsBeforeBattle(getCurrentFighters());
-            battle = true;
-        }
-    }
-
-    protected class CountdownTitleTask extends BukkitRunnable {
-
-        private final Collection<Warrior> warriors;
-        private int timer;
-
-        public CountdownTitleTask(Collection<Warrior> warriors, int timer) {
-            this.warriors = warriors;
-            if (timer < 0) {
-                timer = 0;
-            }
-            this.timer = timer;
-        }
-
-        @SuppressWarnings("deprecation")
-        @Override
-        public void run() {
-            List<Player> players = warriors.stream().map(Warrior::toOnlinePlayer).filter(Objects::nonNull)
-                    .collect(Collectors.toList());
-            String title;
-            if (timer > 0) {
-                title = getColor() + "" + timer;
-            } else {
-                title = ChatColor.RED + plugin.getLang("title.fight", Game.this);
-                this.cancel();
-                Bukkit.getScheduler().runTaskLater(plugin, () -> players.forEach(Player::resetTitle), 20L);
-            }
-            players.forEach(player -> player.sendTitle(title, ""));
-            timer--;
-        }
-
-        private ChatColor getColor() {
-            ChatColor color;
-            if (timer > 6) {
-                color = ChatColor.GREEN;
-            } else if (timer > 3) {
-                color = ChatColor.YELLOW;
-            } else {
-                color = ChatColor.RED;
-            }
-            return color;
-        }
+    @Override
+    protected void killTasks() {
+        super.killTasks();
+        plugin.getTaskManager().killAllTasks();
     }
 
     protected class LobbyAnnouncementTask extends BukkitRunnable {
@@ -548,9 +165,8 @@ public abstract class Game extends BaseGame {
         public void run() {
             seconds = times * interval;
             if (times > 0) {
-                Bukkit.getServer().broadcastMessage(MessageFormat.format(plugin.getLang("starting_game", Game.this),
-                        seconds, getConfig().getMinimumGroups(), getConfig().getMinimumPlayers(),
-                        getGroupParticipants().size(), getParticipants().size()));
+                broadcastKey("starting_game", seconds, getConfig().getMinimumGroups(),
+                        getConfig().getMinimumPlayers(), getGroupParticipants().size(), getParticipants().size());
                 times--;
             } else {
                 preLobbyEnd();
@@ -582,14 +198,4 @@ public abstract class Game extends BaseGame {
         }
     }
 
-    protected class GameExpirationTask extends BukkitRunnable {
-
-        @Override
-        public void run() {
-            gameManager.getCurrentGame().ifPresent(game -> {
-                game.finish(true);
-                gameManager.broadcastKey("game_expired", game);
-            });
-        }
-    }
 }
