+++ src/main/resources/language-en.yml
@@ -1,3 +1,10 @@
+you.must.be.in.a.group.to.challenge: "&c» You must be in a group to challenge another group!"
+challenged.you: "&a» {0} challenged you to a duel! Accept it using /tb challenge accept {0}"
+challenged.your.group: "&a» {0} challenged your group to a battle! Accept it using /tb challenge accept {0}"
+your.group.challenged: "&a» Your group challenged {1}! Join the battle using /tb challenge accept {0}"
+you.challenged.player: "&a» You challenged {0} to a duel!"
+you.challenged.group: "&a» You challenged the group {0} to a battle!"
+already.challenged: "&c» You already challenged someone!"
 teleport.error: "&c» An error ocurred while teleporting you"
 title.fight: "Fight!"
 wait_for_your_turn: "&b» Duel prepared, wait for your turn!"
